+++ DeviceSDK/src/main/java/org/astarteplatform/devicesdk/protocol/AstarteServerPropertyInterface.java
@@ -4,10 +4,14 @@ import java.util.Collection;
 import java.util.HashSet;
 import java.util.Map;
 import org.astarteplatform.devicesdk.AstartePropertyStorage;
+import org.astarteplatform.devicesdk.logging.AstarteSDKLogManager;
+import org.astarteplatform.devicesdk.logging.AstarteSDKLogger;
 import org.joda.time.DateTime;
 
 public class AstarteServerPropertyInterface extends AstartePropertyInterface
     implements AstarteServerValueBuilder, AstarteServerValuePublisher {
+  AstarteSDKLogger logger = AstarteSDKLogManager.INSTANCE.getLogger();
+
   private final Collection<AstartePropertyEventListener> mListeners;
 
   AstarteServerPropertyInterface(AstartePropertyStorage propertyStorage) {
@@ -52,8 +56,9 @@ public class AstarteServerPropertyInterface extends AstartePropertyInterface
               .build();
     } else {
       // Couldn't find the mapping
-      System.err.printf(
-          "Got an unexpected path %s for interface %s!%n", interfacePath, getInterfaceName());
+      logger.error(
+          String.format(
+              "Got an unexpected path %s for interface %s!%n", interfacePath, getInterfaceName()));
 
       astarteServerValue = null;
     }