+++ examples/Generic/src/main/java/org/astarteplatform/devicesdk/generic/examples/ExampleMessageListener.java
@@ -1,14 +1,18 @@
 package org.astarteplatform.devicesdk.generic.examples;
 
 import org.astarteplatform.devicesdk.AstarteMessageListener;
+import org.astarteplatform.devicesdk.logging.AstarteSDKLogManager;
+import org.astarteplatform.devicesdk.logging.AstarteSDKLogger;
 
 public class ExampleMessageListener implements AstarteMessageListener {
+  final AstarteSDKLogger logger = AstarteSDKLogManager.INSTANCE.getLogger();
+
   public void onConnected() {
     /*
      * This function gets called when the device establishes the connection
      * with the broker.
      */
-    System.out.println("Device connected");
+    logger.info("Device connected");
   }
 
   public void onDisconnected(Throwable cause) {
@@ -16,7 +20,7 @@ public class ExampleMessageListener implements AstarteMessageListener {
      * This function gets called when the device loses the connection with the
      * broker.
      */
-    System.out.println("Device disconnected: " + cause.getMessage());
+    logger.info("Device disconnected: " + cause.getMessage());
   }
 
   public void onFailure(Throwable cause) {
@@ -24,6 +28,6 @@ public class ExampleMessageListener implements AstarteMessageListener {
      * This function gets called when the device encounters an error during its
      * lifetime.
      */
-    System.out.println("Device failure: " + cause.getMessage());
+    logger.error("Device failure: " + cause.getMessage());
   }
 }