+++ CHANGELOG.md
@@ -6,11 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [1.1.0-dev] - Unreleased
 ### Added
+- Logging level is now handled using the `org.astarteplatform.devicesdk.logging.level` VM Property
 - Add support to array-type endpoints.
 - Sending data on an endpoint not present in interface now triggers an exception.
 - Sending an object with fewer data than needed now triggers an exception.
 - Sending a non-numeric value on an endpoint with type Double now triggers an exception.
 ### Changed
+- The logs are now handled by a dedicated class
 - Data is now validated against their exact types.
 - Update Gradle to latest stable 7.2
 - [android] Update Android Gradle plugin and Android target version